requirements.txt
google-generativeai==0.8.3
moviepy==1.0.3
gtts==2.5.4
pandas==2.3.3
sqlalchemy==2.0.43
google-api-python-client==2.183.0
google-auth-oauthlib==1.2.2
beautifulsoup4==4.12.3
wikipedia-api==0.7.1
cairosvg==2.7.1
scikit-learn==1.5.2
tenacity==9.0.0
markovify==0.9.4
requests==2.32.5
boto3==1.40.41
pillow==11.3.0pydub==0.25.1
